+++ App/.idea/misc.xml
@@ -10,6 +10,7 @@
         <entry key="..\:/Users/Admin/Desktop/Hackathons/SIH/App/app/src/main/res/layout/activity_signup_final_upload.xml" value="0.1308876811594203" />
         <entry key="..\:/Users/Admin/Desktop/Hackathons/SIH/App/app/src/main/res/layout/activity_signup_newemployee.xml" value="0.1308876811594203" />
         <entry key="..\:/Users/Admin/Desktop/Hackathons/SIH/App/app/src/main/res/layout/activity_signup_newemployee_cred.xml" value="0.12" />
+        <entry key="..\:/Users/Admin/Desktop/Hackathons/SIH/App/app/src/main/res/layout/activity_threeshot.xml" value="0.17552083333333332" />
         <entry key="..\:/Users/Admin/Desktop/Hackathons/SIH/App/app/src/main/res/layout/check_status_layout.xml" value="0.1" />
         <entry key="..\:/Users/Admin/Desktop/Hackathons/SIH/App/app/src/main/res/layout/dashboard.xml" value="0.18025078369905956" />
         <entry key="..\:/Users/Admin/Desktop/Hackathons/SIH/App/app/src/main/res/layout/existing_user.xml" value="0.14130434782608695" />
